oj-agent/pyproject.toml
[project]
name = "oj-agent"
version = "0.1.0"
description = "Standalone LangGraph-based OJ tutor service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "langgraph>=0.2.0",
    "openai>=1.52.0",
    "pydantic>=2.9.0",
    "PyYAML>=6.0.2",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
